# prod/app/api/export-image/route.ts
import { NextRequest, NextResponse } from 'next/server'

/**
 * Image Export API Route
 * 
 * Handles high-quality image export with custom resolutions
 */

interface ExportRequest {
  canvasJson: string
  format: 'png' | 'jpg'
  quality?: number
  width?: number
  height?: number
  multiplier?: number // DPI multiplier for export quality
}

export async function POST(request: NextRequest) {
  try {
    const body: ExportRequest = await request.json()

    if (!body.canvasJson) {
      return NextResponse.json(
        { error: 'Missing canvas data' },
        { status: 400 }
      )
    }

    const {
      format = 'png',
      quality = 0.95,
      multiplier = 2,
    } = body

    console.log('[v0] Image export request:', {
      format,
      quality,
      multiplier,
    })

    // In a real implementation:
    // 1. Parse the canvas JSON
    // 2. Render using fabric.js server-side or node-canvas
    // 3. Apply quality settings
    // 4. Convert to requested format
    // 5. Return as download

    // For MVP, return success confirmation
    return NextResponse.json({
      success: true,
      message: `Image export to ${format.toUpperCase()} with ${multiplier}x quality`,
      format,
      quality,
      multiplier,
      timestamp: new Date().toISOString(),
    })
  } catch (error) {
    console.error('[v0] Image export error:', error)
    return NextResponse.json(
      { error: 'Export failed' },
      { status: 500 }
    )
  }
}